requirements.txt
torch==2.1.0
torchvision==0.16.1
pillow
opencv-python
streamlit
